fitur materi, kuis, presensi

+++ app/Http/Controllers/kelasController.php
@@ -5,12 +5,19 @@ namespace App\Http\Controllers;
 use App\Models\kelas;
 use App\Models\siswa;
 use Illuminate\Http\Request;
+use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\DB;
 
 class kelasController extends Controller
 {
     //
     public function add_kelas(Request $request){
+        $role = session("role");
+        $email = session("email");
+        
+        if($role != "admin"){
+            return abort(403);
+        }
         $kelas = $request->kelas;
         $kuota = $request->kuota;
         
@@ -25,7 +32,36 @@ class kelasController extends Controller
             "kuota_siswa" => $kuota,
         ]);
         if($insertKelas == true){
-            return back()->with(['sukses_add' => "Berhasil menambah kelas"]);
+            //get id kelas yang udah ditambah barusan
+            $getIdKelas = kelas::where("nama_kelas",'=',$kelas)->first();
+            //menambah data jadwal selama setahun dari kelas ditambahkan
+            if ($getIdKelas == true) {
+                $id_kelas = $getIdKelas->id_kelas;
+                $startDate = Carbon::now(); // Tanggal mulai (hari ini)
+                $endDate = Carbon::now()->addYear(); // Tanggal berakhir (1 tahun ke depan)
+            
+                $currentDate = $startDate;
+                $data = []; // Array untuk menyimpan data batch insert
+            
+                while ($currentDate->lte($endDate)) {
+                    // Cek apakah hari ini adalah Senin-Jumat
+                    if ($currentDate->isWeekday()) {
+                        $data[] = [
+                            'id_jadwal' => 'JDWL' . $currentDate->format('y') . $currentDate->format('m') . $currentDate->format('d'). $id_kelas,
+                            'id_kelas' => $id_kelas,
+                            'tanggal_pelaksanaan' => $currentDate->format('Y-m-d'),
+                            'status' => 'aktif', // Ubah status sesuai kebutuhan
+                        ];
+                    }
+                    $currentDate->addDay(); // Tambahkan 1 hari
+                }
+                $insertJadwal = DB::table('jadwals')->insert($data);
+                if($insertJadwal == true){
+                    return back()->with(['sukses_add' => "Berhasil menambah kelas"]);
+                }else{
+                    return back()->with(['error_add' => "Gagal menambah kelas"]);
+                }
+            }
         }else{
             return back()->with(['error_add' => "Gagal menambah kelas"]);
         }
@@ -63,15 +99,27 @@ class kelasController extends Controller
     }
 
     public function delete_kelas($id){
+        $role = session("role");
+        $email = session("email");
+        
+        if($role != "admin"){
+            return abort(403);
+        }
         // cek apakah ada siswa yang berada di kelas ini
         $cekSiswa = siswa::where("id_kelas",'=',$id)->first();
         if($cekSiswa == true){
             return back()->with(['error_delete' => "Terdapat Siswa"]);
         }
 
-        $delete = DB::table('kelas')->where("id_kelas",'=',$id)->delete();
-        if($delete == true){
-            return back()->with(['sukses_delete' => "Berhasil menghapus kelas"]);
+        //menghapus jadwal terlebih dahulu
+        $deleteJadwal = DB::table("jadwals")->where("id_kelas",'=',$id)->delete();
+        if($deleteJadwal == true){
+            $delete = DB::table('kelas')->where("id_kelas",'=',$id)->delete();
+            if($delete == true){
+                return back()->with(['sukses_delete' => "Berhasil menghapus kelas"]);
+            }else{
+                return back()->with(['error_delete' => "Berhasil menghapus kelas"]);
+            }
         }else{
             return back()->with(['error_delete' => "Berhasil menghapus kelas"]);
         }
